LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/input-valueasdate-expected.txt

input_valueAsDate


PASS valueAsDate getter on type date (with value: "")
PASS valueAsDate getter on type date (with value: "0000-12-10")
PASS valueAsDate getter on type date (with value: "2019-00-12")
PASS valueAsDate getter on type date (with value: "2019-12-00")
PASS valueAsDate getter on type date (with value: "2019-13-10")
PASS valueAsDate getter on type date (with value: "2019-02-29")
FAIL valueAsDate getter on type date (with value: "2019-12-10") assert_equals: expected object "Mon Dec 09 2019 16:00:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate getter on type date (with value: "2016-02-29") assert_equals: expected object "Sun Feb 28 2016 16:00:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate setter on type date (new Date("2019-12-10T00:00:00.000Z")) The object is in an invalid state.
FAIL valueAsDate setter on type date (new Date("2016-02-29T00:00:00.000Z")) The object is in an invalid state.
PASS valueAsDate getter on type month (with value: "")
PASS valueAsDate getter on type month (with value: "0000-12")
PASS valueAsDate getter on type month (with value: "2019-00")
FAIL valueAsDate getter on type month (with value: "2019-12") assert_equals: expected object "Sat Nov 30 2019 16:00:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate setter on type month (new Date("2019-12-01T00:00:00.000Z")) The object is in an invalid state.
PASS valueAsDate getter on type week (with value: "")
PASS valueAsDate getter on type week (with value: "0000-W50")
PASS valueAsDate getter on type week (with value: "2019-W00")
PASS valueAsDate getter on type week (with value: "2019-W60")
FAIL valueAsDate getter on type week (with value: "2019-W50") assert_equals: expected object "Sun Dec 08 2019 16:00:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate setter on type week (new Date("2019-12-09T00:00:00.000Z")) The object is in an invalid state.
PASS valueAsDate getter on type time (with value: "")
PASS valueAsDate getter on type time (with value: "24:00")
PASS valueAsDate getter on type time (with value: "00:60")
FAIL valueAsDate getter on type time (with value: "00:00") assert_equals: expected object "Wed Dec 31 1969 16:00:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate getter on type time (with value: "12:00") assert_equals: expected object "Thu Jan 01 1970 04:00:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate getter on type time (with value: "23:59") assert_equals: expected object "Thu Jan 01 1970 15:59:00 GMT-0800 (Pacific Standard Time)" but got null
FAIL valueAsDate setter on type time (new Date("1970-01-01T00:00:00.000Z")) The object is in an invalid state.
FAIL valueAsDate setter on type time (new Date("1970-01-01T12:00:00.000Z")) The object is in an invalid state.
FAIL valueAsDate setter on type time (new Date("1970-01-01T23:59:00.000Z")) The object is in an invalid state.
